Improve Transporter path finding performance and accuracy (#5846)
* Start work on trying to optimize path finding some and fix various minor bugs with it such as #5843

* Work on optimizing some more things about transporter paths to reduce the number of unneeded getTileEntity calls made, and to do more efficient ones where possible

* Use a single chunkMap/cache for each TransporterPathfinder.getPaths call, as the various chunks in it are likely to be accessed by all the different parts of it. Also use a chunk map cache for minorly optimizing networks output location finding

* Fix issue with exiting early just because there is an arbitrary path that quickly leaves the max check distance. Also inline getEstimate so it is more clear how it works

* Slightly improve potential performance of calculating the idle path

* Remove a couple TODOs and improve on logic for checking if we have a valid direction. We now do a complete check if we are directly next to our destination, and exit early with the proper path if we are and it is a valid path. Rather than having to loop through to find it

* Cleanup another TODO using an already calculated TileEntity where we have one for checking inserting into a transporter

* Inline a couple methods from TransporterUtils to make it easier to tell what is going on

* Minor cleanup to tile's querying the world to get themselves, and add a comment to TransporterImpl

* Replace a few implementations of collections in TransporterPathfinder with FastUtil implementations

+++ src/api/java/mekanism/api/Coord4D.java
@@ -233,11 +233,11 @@ public class Coord4D {//TODO: Replace this with GlobalPos
      *
      * @return the distance to the defined Coord4D
      */
-    public int distanceTo(Coord4D obj) {
+    public double distanceTo(Coord4D obj) {
         int subX = x - obj.x;
         int subY = y - obj.y;
         int subZ = z - obj.z;
-        return (int) MathHelper.sqrt(subX * subX + subY * subY + subZ * subZ);
+        return MathHelper.sqrt(subX * subX + subY * subY + subZ * subZ);
     }
 
     /**
